update GDAL to 3.13 (#349)
* update GDAL to 3.13

* add GDT_UInt8 as a named constant

* disable this test for now

* try enabling all SIMD instruction sets

* fix concatenated macros

* enable AVX/AVX2 on the compiler level

* enable AVX for macOS clang

* enable via XCode and only on x86

* show the command line when building

* is it x64?

* prefer std::min to MIN

* enable SSE/AVX only on x86

* enable 64 bit file access on Linux

* enable NEON SIMD

* fix a typo

* drop NEON for now

* run all network tests

* remove SIMD from the default build

* this depends on the GDAL version and can be cached

* update the changelog

* actually test what the test title says

* fix compilation with MSVC and without AVX2

+++ deps/libgdal/gdal/ogr/swq_op_general.cpp
@@ -166,7 +166,7 @@ int swq_test_like(const char *input, const char *pattern, char chEscape,
 }
 
 /************************************************************************/
-/*                        OGRHStoreGetValue()                           */
+/*                         OGRHStoreGetValue()                          */
 /************************************************************************/
 
 static char *OGRHStoreCheckEnd(char *pszIter, int bIsKey)
@@ -319,7 +319,7 @@ char *OGRHStoreGetValue(const char *pszHStore, const char *pszSearchedKey)
 
 #ifdef DEBUG_VERBOSE
 /************************************************************************/
-/*                         OGRFormatDate()                              */
+/*                           OGRFormatDate()                            */
 /************************************************************************/
 
 #ifdef __GNUC__
@@ -331,7 +331,7 @@ static const char *OGRFormatDate(const OGRField *psField)
     return CPLSPrintf("%04d/%02d/%02d %02d:%02d:%06.3f", psField->Date.Year,
                       psField->Date.Month, psField->Date.Day,
                       psField->Date.Hour, psField->Date.Minute,
-                      psField->Date.Second);
+                      static_cast<double>(psField->Date.Second));
 }
 
 #ifdef __GNUC__
@@ -1229,7 +1229,7 @@ swq_expr_node *SWQGeneralEvaluator(swq_expr_node *node,
 }
 
 /************************************************************************/
-/*                SWQAutoPromoteIntegerToInteger64OrFloat()             */
+/*              SWQAutoPromoteIntegerToInteger64OrFloat()               */
 /************************************************************************/
 
 static void SWQAutoPromoteIntegerToInteger64OrFloat(swq_expr_node *poNode)
@@ -1278,7 +1278,7 @@ static void SWQAutoPromoteIntegerToInteger64OrFloat(swq_expr_node *poNode)
 }
 
 /************************************************************************/
-/*                    SWQAutoPromoteStringToDateTime()                  */
+/*                   SWQAutoPromoteStringToDateTime()                   */
 /************************************************************************/
 
 static void SWQAutoPromoteStringToDateTime(swq_expr_node *poNode)
@@ -1378,7 +1378,7 @@ static void SWQAutoConvertStringToNumeric(swq_expr_node *poNode)
 }
 
 /************************************************************************/
-/*                   SWQCheckSubExprAreNotGeometries()                  */
+/*                  SWQCheckSubExprAreNotGeometries()                   */
 /************************************************************************/
 
 static bool SWQCheckSubExprAreNotGeometries(swq_expr_node *poNode)
